+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Added `Debug` trait bound on `Key`
 - Impl `Display` for `Error`
+- Added mutation tracking.
+  - Items are now only flushed they are newly added, or when mutations occur.
+  - This change should be transparent to the users of this crate.
 
 ## v1.2.1